More portable file management. Verify on linux and check Wii's un-save-able games.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@1405 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Core/Src/IPC_HLE/WII_IPC_HLE_Device_FileIO.cpp
@@ -216,7 +216,7 @@ CWII_IPC_HLE_Device_FileIO::IOCtl(u32 _CommandAddress)
             LOG(WII_IPC_FILEIO, "FileIO: ISFS_IOCTL_GETFILESTATS");
             LOG(WII_IPC_FILEIO, "    Length: %i   Seek: %i", m_FileLength, Position);
 
-            Memory::Write_U32(m_FileLength, BufferOut);
+            Memory::Write_U32((u32)m_FileLength, BufferOut);
             Memory::Write_U32(Position, BufferOut+4);
         }
         break;
