+++ Source/Core/Core/Src/IPC_HLE/WII_IPC_HLE_Device_FileIO.h
@@ -65,7 +65,7 @@ private:
     };
 
     FILE* m_pFileHandle;
-    u32 m_FileLength;
+    u64 m_FileLength;
 
 	std::string m_Filename;
 };
